+++ Sources/Buffie/File Writers/HLS/Playlists/HLSPlaylistType.swift
@@ -2,7 +2,7 @@ import Foundation
 
 protocol PlaylistWriter {
     func positionToSeek() -> UInt64?
-    func header(with targetDuration: Int64) -> String
+    func header(with targetDuration: Double) -> String
     func writeSegment(with filename: String, duration: Float64, and firstMediaSequence: Int) -> String
     func end() -> String
 }
